backend/requirements.txt
flask==3.0.0
google-cloud-aiplatform==1.47.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-tasks==2.16.0
google-auth==2.27.0
moviepy==1.0.3
Pillow==10.2.0
